Make generic 3D viewer class (with turntable controls) for future 3d viz, have visualizer inherit from that insted of re-rolling code every time
