+++ apps/visualizer/app.py
@@ -9,7 +9,7 @@ import vtkmodules.all as vtk
 
 # local (visualizer app & apps.shared)
 from apps.shared.screen_management import set_up_render_window_bounds
-from apps.shared.app import App
+from apps.shared.generic_3d_viewer_app import Generic3DViewerApp
 from apps.visualizer import utilities
 from apps.visualizer.mesh import Mesh, MeshColorMode
 from apps.visualizer.point_cloud import PointCloud, PointColorMode
@@ -17,9 +17,10 @@ from apps.visualizer.correspondence_line_set import CorrespondenceColorMode, Cor
 from apps.visualizer.graph import Graph
 
 
-class VisualizerApp(App):
+class VisualizerApp(Generic3DViewerApp):
 
     def __init__(self, output_path: Path, start_frame_ix: int = -1, outgoing_queue: Union[None, Queue] = None):
+        super().__init__()
         self.__alt_pressed = False
         self.outgoing_queue = outgoing_queue
 
@@ -31,19 +32,11 @@ class VisualizerApp(App):
             raise ValueError(f"Smallest start frame for given sequence is {minimum_start_frame:d}, "
                              f"cannot start from frame {start_frame_ix:d}.")
 
-        # self.inverse_camera_matrices = trajectory_loading.load_inverse_matrices(output_path)
         self.start_frame_ix = start_frame_ix
         self.output_path = output_path
-        self.offset_cam = (0.2562770766576, 0.13962609403401335, -0.2113334598208764)
-        colors = vtk.vtkNamedColors()
         self.current_frame = start_frame_ix
 
-        # renderer & render window initialization
-        self.renderer = vtk.vtkRenderer()
-        self.render_window = vtk.vtkRenderWindow()
-        self.render_window.SetWindowName("Visualizer: " + str(output_path))
-        set_up_render_window_bounds(self.render_window, None, 1)
-        self.render_window.AddRenderer(self.renderer)
+        colors = vtk.vtkNamedColors()
 
         # mesh setup
         self.canonical_mesh = Mesh(self.renderer, self.render_window, colors.GetColor3d("Peacock"))
@@ -91,7 +84,7 @@ class VisualizerApp(App):
         if self.have_graph_info:
             self.graph = Graph(self.renderer, self.render_window, colors.GetColor3d("Green"))
 
-        # text setup
+        # === text setup ===
         self.text_mapper = vtk.vtkTextMapper()
         shown_point_cloud = "none" if len(self.point_cloud_names) == 0 \
             else self.point_cloud_names[self.shown_point_cloud_index]
@@ -109,58 +102,19 @@ class VisualizerApp(App):
         self.font_size = 20
         text_property.SetFontSize(self.font_size)
         text_property.SetColor(colors.GetColor3d('Mint'))
-
+        # text actor
         self.text_actor = vtk.vtkActor2D()
         self.text_actor.SetMapper(self.text_mapper)
         self.last_window_width, self.last_window_height = 0, 0
         self.renderer.AddActor(self.text_actor)
 
-        # background
-        self.renderer.SetBackground(colors.GetColor3d("Black"))
-
-        # Interactor setup
-        self.interactor = vtk.vtkRenderWindowInteractor()
-        self.interactor.SetInteractorStyle(None)
-        self.interactor.SetRenderWindow(self.render_window)
-        self.interactor.Initialize()
-
-        self.interactor.AddObserver("ModifiedEvent", self.update_window)
-        self.interactor.AddObserver("KeyPressEvent", self.key_press)
-        self.interactor.AddObserver("KeyReleaseEvent", self.key_release)
-        self.interactor.AddObserver("LeftButtonPressEvent", self.button_event)
-        self.interactor.AddObserver("LeftButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("RightButtonPressEvent", self.button_event)
-        self.interactor.AddObserver("RightButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("MiddleButtonPressEvent", self.button_event)
-        self.interactor.AddObserver("MiddleButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("MiddleButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("MouseWheelForwardEvent", self.button_event)
-        self.interactor.AddObserver("MouseWheelBackwardEvent", self.button_event)
-        self.interactor.AddObserver("MouseMoveEvent", self.mouse_move)
-
-        self.rotating = False
-        self.panning = False
-        self.zooming = False
-
-        # axes actor
-        self.axes = vtk.vtkAxesActor()
-        self.axes_widget = widget = vtk.vtkOrientationMarkerWidget()
-        rgba = colors.GetColor4ub("Carrot")
-        widget.SetOutlineColor(rgba[0], rgba[1], rgba[2])
-        widget.SetOrientationMarker(self.axes)
-        widget.SetInteractor(self.interactor)
-        widget.SetViewport(0.0, 0.2, 0.2, 0.4)
-        widget.SetEnabled(1)
-        widget.InteractiveOn()
-
-        self.camera = camera = self.renderer.GetActiveCamera()
-
         # set up camera
-        camera.SetPosition(self.offset_cam[0], self.offset_cam[1], self.offset_cam[2])
-        camera.SetPosition(0, 0, -1)
-        camera.SetViewUp(0, 1.0, 0)
-        camera.SetFocalPoint(0, 0, 1.5)
-        camera.SetClippingRange(0.01, 10.0)
+        self.offset_cam = (0.2562770766576, 0.13962609403401335, -0.2113334598208764)
+        self.camera.SetPosition(self.offset_cam[0], self.offset_cam[1], self.offset_cam[2])
+        self.camera.SetPosition(0, 0, -1)
+        self.camera.SetViewUp(0, 1.0, 0)
+        self.camera.SetFocalPoint(0, 0, 1.5)
+        self.camera.SetClippingRange(0.01, 10.0)
 
         self.update_window(None, None)
         self.set_frame(self.current_frame)
@@ -177,9 +131,8 @@ class VisualizerApp(App):
             self.correspondence_set.hide()
 
         self.update_text()
-        self.render_window.Render()
-
         self._pixel_labels_visible = False
+        self.render_window.Render()
 
     def launch(self):
         # Start the event loop.
@@ -307,118 +260,6 @@ class VisualizerApp(App):
             else:
                 self.show_point_cloud_at_index(self.shown_point_cloud_index - 1)
 
-    # Handle the mouse button events.
-    def button_event(self, obj, event):
-        if event == "LeftButtonPressEvent":
-            self.rotating = True
-        elif event == "LeftButtonReleaseEvent":
-            self.rotating = False
-        elif event == "RightButtonPressEvent":
-            self.panning = True
-        elif event == "RightButtonReleaseEvent":
-            self.panning = False
-        elif event == "MiddleButtonPressEvent":
-            self.zooming = True
-        elif event == "MiddleButtonReleaseEvent":
-            self.zooming = False
-        elif event == "MouseWheelForwardEvent":
-            self.dolly_step(1)
-        elif event == "MouseWheelBackwardEvent":
-            self.dolly_step(-1)
-
-    def mouse_move(self, obj, event):
-        last_x_y_pos = self.interactor.GetLastEventPosition()
-        last_x = last_x_y_pos[0]
-        last_y = last_x_y_pos[1]
-
-        x_y_pos = self.interactor.GetEventPosition()
-        x = x_y_pos[0]
-        y = x_y_pos[1]
-
-        center = self.render_window.GetSize()
-        center_x = center[0] / 2.0
-        center_y = center[1] / 2.0
-
-        if self.rotating:
-            self.rotate(x, y, last_x, last_y)
-        elif self.panning:
-            self.pan(x, y, last_x, last_y, center_x, center_y)
-        elif self.zooming:
-            self.dolly(x, y, last_x, last_y)
-
-    def rotate(self, x, y, last_x, last_y):
-        speed = 0.5
-        self.camera.Azimuth(speed * (last_x - x))
-        self.camera.Elevation(speed * (last_y - y))
-        self.camera.SetViewUp(0, 0, 0)
-        self.render_window.Render()
-
-    def pan(self, x, y, last_x, last_y, center_x, center_y):
-        renderer = self.renderer
-        camera = self.camera
-        f_point = camera.GetFocalPoint()
-        f_point0 = f_point[0]
-        f_point1 = f_point[1]
-        f_point2 = f_point[2]
-
-        p_point = camera.GetPosition()
-        p_point0 = p_point[0]
-        p_point1 = p_point[1]
-        p_point2 = p_point[2]
-
-        renderer.SetWorldPoint(f_point0, f_point1, f_point2, 1.0)
-        renderer.WorldToDisplay()
-        d_point = renderer.GetDisplayPoint()
-        focal_depth = d_point[2]
-
-        a_point0 = center_x + (x - last_x)
-        a_point1 = center_y + (y - last_y)
-
-        renderer.SetDisplayPoint(a_point0, a_point1, focal_depth)
-        renderer.DisplayToWorld()
-        r_point = renderer.GetWorldPoint()
-        r_point0 = r_point[0]
-        r_point1 = r_point[1]
-        r_point2 = r_point[2]
-        r_point3 = r_point[3]
-
-        if r_point3 != 0.0:
-            r_point0 = r_point0 / r_point3
-            r_point1 = r_point1 / r_point3
-            r_point2 = r_point2 / r_point3
-
-        camera.SetFocalPoint((f_point0 - r_point0) / 2.0 + f_point0,
-                             (f_point1 - r_point1) / 2.0 + f_point1,
-                             (f_point2 - r_point2) / 2.0 + f_point2)
-        camera.SetPosition((f_point0 - r_point0) / 2.0 + p_point0,
-                           (f_point1 - r_point1) / 2.0 + p_point1,
-                           (f_point2 - r_point2) / 2.0 + p_point2)
-        self.render_window.Render()
-
-    def dolly(self, x, y, last_x, last_y):
-        dolly_factor = pow(1.02, (0.5 * (y - last_y)))
-        camera = self.camera
-        if camera.GetParallelProjection():
-            parallel_scale = camera.GetParallelScale() * dolly_factor
-            camera.SetParallelScale(parallel_scale)
-        else:
-            camera.Dolly(dolly_factor)
-            self.renderer.ResetCameraClippingRange()
-
-        self.render_window.Render()
-
-    def dolly_step(self, step):
-        dolly_factor = pow(1.02, (10.0 * step))
-        camera = self.camera
-        if camera.GetParallelProjection():
-            parallel_scale = camera.GetParallelScale() * dolly_factor
-            camera.SetParallelScale(parallel_scale)
-        else:
-            camera.Dolly(dolly_factor)
-            self.renderer.ResetCameraClippingRange()
-
-        self.render_window.Render()
-
     KEYS_TO_SEND = {"q", "Escape", "bracketright", "bracketleft"}
 
     def key_press(self, obj, event):
@@ -532,5 +373,3 @@ class VisualizerApp(App):
                                   f"Point color mode: {self.point_color_mode.name:s}\n"
                                   f"Visible correspondences: {visible_correspondence_percentage}%\n"
                                   f"Corresp. color mode: {self.correspondence_color_mode.name:s}\n")
-
-
