+++ subprocedure_examples/visualize_blocks.py
@@ -1,193 +1,7 @@
-# system
-import sys
-# third-party
-import pyglet
-import vtkmodules.all as vtk
+from apps.shared.generic_3d_viewer_app import Generic3DViewerApp
 
 
-
-
-
-
-class GenericVisualizerApp:
-    def __init__(self):
-        colors = vtk.vtkNamedColors()
-
-        # renderer & render window initialization
-        self.renderer = vtk.vtkRenderer()
-        self.render_window = vtk.vtkRenderWindow()
-        self.render_window.SetWindowName("Blocks")
-        display = pyglet.canvas.get_display()
-        screens = display.get_screens()
-        screens.sort(key=lambda screen: (screen.x, screen.y))
-        chosen_screen_index = 0
-        chosen_screen = screens[chosen_screen_index]
-        self.render_window.SetPosition(chosen_screen.x, chosen_screen.y)
-        self.render_window.SetSize((chosen_screen.width, chosen_screen.height))
-        self.render_window.AddRenderer(self.renderer)
-        self.renderer.SetBackground(colors.GetColor3d("Black"))
-
-        # Interactor setup
-        self.interactor = vtk.vtkRenderWindowInteractor()
-        self.interactor.SetInteractorStyle(None)
-        self.interactor.SetRenderWindow(self.render_window)
-        self.interactor.Initialize()
-        self.interactor.AddObserver("KeyPressEvent", self.key_press)
-        self.interactor.AddObserver("LeftButtonPressEvent", self.button_event)
-        self.interactor.AddObserver("LeftButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("RightButtonPressEvent", self.button_event)
-        self.interactor.AddObserver("RightButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("MiddleButtonPressEvent", self.button_event)
-        self.interactor.AddObserver("MiddleButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("MiddleButtonReleaseEvent", self.button_event)
-        self.interactor.AddObserver("MouseWheelForwardEvent", self.button_event)
-        self.interactor.AddObserver("MouseWheelBackwardEvent", self.button_event)
-        self.interactor.AddObserver("MouseMoveEvent", self.mouse_move)
-
-        # axes actor
-        self.axes = vtk.vtkAxesActor()
-        self.axes_widget = widget = vtk.vtkOrientationMarkerWidget()
-        rgba = colors.GetColor4ub("Carrot")
-        widget.SetOutlineColor(rgba[0], rgba[1], rgba[2])
-        widget.SetOrientationMarker(self.axes)
-        widget.SetInteractor(self.interactor)
-        widget.SetViewport(0.0, 0.2, 0.2, 0.4)
-        widget.SetEnabled(1)
-        widget.InteractiveOn()
-
-        self.camera = self.renderer.GetActiveCamera()
-
-        self.rotating = False
-        self.panning = False
-        self.zooming = False
-
-        self.render_window.Render()
-
-    def launch(self):
-        # Start the event loop.
-        self.interactor.Start()
-
-    def key_press(self, obj, event):
-        key = obj.GetKeySym()
-        print("Key:", key)
-        # ==== window controls ===========
-        if key == "q" or key == "Escape":
-            self.interactor.InvokeEvent("DeleteAllObjects")
-            sys.exit()
-
-    def button_event(self, obj, event):
-        if event == "LeftButtonPressEvent":
-            self.rotating = True
-        elif event == "LeftButtonReleaseEvent":
-            self.rotating = False
-        elif event == "RightButtonPressEvent":
-            self.panning = True
-        elif event == "RightButtonReleaseEvent":
-            self.panning = False
-        elif event == "MiddleButtonPressEvent":
-            self.zooming = True
-        elif event == "MiddleButtonReleaseEvent":
-            self.zooming = False
-        elif event == "MouseWheelForwardEvent":
-            self.dolly_step(1)
-        elif event == "MouseWheelBackwardEvent":
-            self.dolly_step(-1)
-
-    def rotate(self, x, y, last_x, last_y):
-        speed = 0.5
-        self.camera.Azimuth(speed * (last_x - x))
-        self.camera.Elevation(speed * (last_y - y))
-        self.camera.SetViewUp(0, 0, 0)
-        self.render_window.Render()
-
-    def pan(self, x, y, last_x, last_y, center_x, center_y):
-        renderer = self.renderer
-        camera = self.camera
-        f_point = camera.GetFocalPoint()
-        f_point0 = f_point[0]
-        f_point1 = f_point[1]
-        f_point2 = f_point[2]
-
-        p_point = camera.GetPosition()
-        p_point0 = p_point[0]
-        p_point1 = p_point[1]
-        p_point2 = p_point[2]
-
-        renderer.SetWorldPoint(f_point0, f_point1, f_point2, 1.0)
-        renderer.WorldToDisplay()
-        d_point = renderer.GetDisplayPoint()
-        focal_depth = d_point[2]
-
-        a_point0 = center_x + (x - last_x)
-        a_point1 = center_y + (y - last_y)
-
-        renderer.SetDisplayPoint(a_point0, a_point1, focal_depth)
-        renderer.DisplayToWorld()
-        r_point = renderer.GetWorldPoint()
-        r_point0 = r_point[0]
-        r_point1 = r_point[1]
-        r_point2 = r_point[2]
-        r_point3 = r_point[3]
-
-        if r_point3 != 0.0:
-            r_point0 = r_point0 / r_point3
-            r_point1 = r_point1 / r_point3
-            r_point2 = r_point2 / r_point3
-
-        camera.SetFocalPoint((f_point0 - r_point0) / 2.0 + f_point0,
-                             (f_point1 - r_point1) / 2.0 + f_point1,
-                             (f_point2 - r_point2) / 2.0 + f_point2)
-        camera.SetPosition((f_point0 - r_point0) / 2.0 + p_point0,
-                           (f_point1 - r_point1) / 2.0 + p_point1,
-                           (f_point2 - r_point2) / 2.0 + p_point2)
-        self.render_window.Render()
-
-    def dolly(self, x, y, last_x, last_y):
-        dolly_factor = pow(1.02, (0.5 * (y - last_y)))
-        camera = self.camera
-        if camera.GetParallelProjection():
-            parallel_scale = camera.GetParallelScale() * dolly_factor
-            camera.SetParallelScale(parallel_scale)
-        else:
-            camera.Dolly(dolly_factor)
-            self.renderer.ResetCameraClippingRange()
-
-        self.render_window.Render()
-
-    def dolly_step(self, step):
-        dolly_factor = pow(1.02, (10.0 * step))
-        camera = self.camera
-        if camera.GetParallelProjection():
-            parallel_scale = camera.GetParallelScale() * dolly_factor
-            camera.SetParallelScale(parallel_scale)
-        else:
-            camera.Dolly(dolly_factor)
-            self.renderer.ResetCameraClippingRange()
-
-        self.render_window.Render()
-
-    def mouse_move(self, obj, event):
-        last_x_y_pos = self.interactor.GetLastEventPosition()
-        last_x = last_x_y_pos[0]
-        last_y = last_x_y_pos[1]
-
-        x_y_pos = self.interactor.GetEventPosition()
-        x = x_y_pos[0]
-        y = x_y_pos[1]
-
-        center = self.render_window.GetSize()
-        center_x = center[0] / 2.0
-        center_y = center[1] / 2.0
-
-        if self.rotating:
-            self.rotate(x, y, last_x, last_y)
-        elif self.panning:
-            self.pan(x, y, last_x, last_y, center_x, center_y)
-        elif self.zooming:
-            self.dolly(x, y, last_x, last_y)
-
-
-class BlockVisualizerApp(GenericVisualizerApp):
+class BlockVisualizerApp(Generic3DViewerApp):
     def __init__(self):
         super.__init__()
         # set up camera
@@ -202,7 +16,7 @@ class BlockVisualizerApp(GenericVisualizerApp):
 
 
 def visualize_blocks():
-    app = GenericVisualizerApp()
+    app = Generic3DViewerApp()
     app.launch()
